Aggregated test cases into their respective files; disabled tests in `TestClientQueries` test case, need refactor to use mock; TODOs marked
