--- tests/integration/__init__.py
@@ -1,11 +1,0 @@
-# Client testing constants
-ISO_DATE = '1999-06-28'
-ISO_START_DATE = ISO_DATE
-ISO_END_DATE = '1999-06-29'
-
-APOD_BY_DATE_QUERY = 'apodByDate'
-APODS_BY_DATE_QUERY = 'apodsByDate'
-APODS_BY_COUNT_QUERY = 'randomApods'
-
-# Assert messages
-RESULT_NO_MATCH = 'The expected result does not match the tested result.'
