--- tests/unit/__init__.py
@@ -1,14 +1,0 @@
-# Client testing constants
-DATE_ARG_SLASH_FMT_ONE = '06/28/1999'
-DATE_ARG_SLASH_FMT_TWO = '28/06/1999'
-DATE_ARG_DOT_FMT_ONE = '06.28.1999'
-DATE_ARG_DOT_FMT_TWO = '28.06.1999'
-DATE_ARG_DASH_FMT_ONE = '06-28-1999'
-DATE_ARG_DASH_FMT_TWO = '28-06-1999'
-
-ISO_DATE_DASH_FMT = '1999-06-28'
-BAD_DATE_FMT = '1999--06--28'
-
-# Assert messages
-OBJ_NO_MATCH = 'The two objects are not of the same instance.'
-DATE_NO_MATCH = 'The tested date does not match the expected date.'
